requirements.txt
python-dotenv==1.0.1
selenium==4.23.1
webdriver-manager==4.0.2 